Add ValueError exception for some operations
Added ValueError exception for the instance enumerate,references,
assocator subcommands because there is the possibility that these
operation call the traditional pywbem operations and if FilterQuery or
FilterQueryLanguage have values and --use-pull-operations is False, the
operation fails in pywbem with a ValueError (i.e. these parameters not
allowed on the EnumerateInstances).

The change converts the ValueError to a click exception and adds tests
for this exception.

This also cleans up the text in two of the click error messages to put
the data in quotes.