Modify the short form values of some general options.
Change the short for --name from -N to -n

Change the short for --noverify from -n to -N

This is because in working with pywbemcli I find that I use the --name
much more often than the --noverify because I put definitions into the
connections file and then reuse the definition.

I had considered changing --name to an argument but found that there are
issues in click with general arguments and command groups that I did not
want to sort out now.

Add short from for --use-pull-ops as -U. -u already used for --user

Modified the test for --timeout to clarify that it is client timeout

Make same modifications to connection add options.