Refine tests for conflicting global parameters (#229)
* Correct processing of globals

1. Insure that the --serve and --mock-server cannot be used
simultaneously.

2. Issue warning if --name is used to name an existing configuration and
other parameters are included noting that they are ignored.

* Add tests to separate --mock-server and --server. Add unit tests

1. Added tests to disallow simultaneous use of --mock-server and
--server.

2. Added tests to warn user if other options are included in pywbemcli
call when the --name options is used. If the connection data is
retrieved from the connection file, we do not allow modification with
other parameters.  The reason for this is that we do not have a
complete picture of the other parameters.  While it would seem logical
to allow overlay of the parameters from the connection file with
modifications, it is not really that straight forward.  Thus, for
example, the --noverify parameter. The user only specifies it when the
True value is desired.  But, at the point we would be modifying existing
parameters in the named connection, we do not really know if the user
wanted to modify the value by the existence of the parameter. It is easy
to determine that many of the parameters would be modified simply by
their existence (i.e. if --timeout is not None, modify the value).
However the -noverify is boolean so only exists if the user wants True,
not if the user wants to use it.  The simplest for now is to not allow
the modifications.

3. Added tests to global tests for the above.

4. Moved some common things to new conftest.py file.

* Add tests/unit/conftest.py to git