fixes #231 issue with exception when server not specified.
Correct error causing accidental exception when a command that calles
the WBEM server is executed with no server defined (no --server or
--mock-server or --name)

This moves the resolution of the server parameters to as late in the
process as possible so that the server (and mock) options do not get
processed into a WBEMCliConnection until that connection is needed. Thus
commands like connection ... and help which do not need the connection
are executed without processing it.

This, however, causes another issue in that we also use the
--mock-server to execute some mocks on pywbemtools itself for our tests
(mock prompt, mock confirm) and this means that these definitons do not
get processed on startup. There are two ways around this:

1. This commit added a hidden general option that sets the mock values.

2. The other alternative is to set a flag in mock files so that the
server can recognize them and process them immediatly on startup, to
define the pywbemtools mocks.