Fix issue #231 unhandled exception on startup
Alternate to another PR that fixes issue # 231.

Fixes the issue of unhandled exception when no --server or --mock-server
by adding test for this condition.  At the same time we moved the code
that creates the WBEMConnection or FakedWBEMConnection to as late as
possible so that only a subcommand that requires the connection activates
the creation.

However fixing the unhandled exception issue  creates a second
problem in that we are using --mock-server
to implement mocks within pywbemcli for tests (mock prompts and mock
confirms) where console input is required. Moving the WBEMConnection
creation causes this pywbemcli test code to not be executed on startup
so that the mocks for the tests are not created.  Thus, we added
a flag in the first line of the files to be included in --mock-server
that tells pywbemcli to eval that file on startup and not put it into
the --mock-server file list (i.e. it is our test, not the users mock).
Note that we consider this an undocumented test feature.

Also removed an artifical call hiearchy in _common (pywbmecliprompt) and
moved from using prompt_toolkit. prompt to click.prompt.

Added tests for the new conditions.