Change name of live_unit directory to to manual
Changed the name of the test live_unit to manual to match pywbem.

Also changed the name of one test file test_global_options.py to
test_general_options.py.  It was my mistake that I started using the
word global for these options early in the project but the help files,
and now all the documentation refer to them as general options.