Put -(dash) in multiword option names
Change a number of the multiword arguments to include - between the
words. This includes the following:  include-classorigin local-only
assoc-class result-class result-role.

Changed the options in centralinstances and changed the command
name to get-centralinsts.

Changed a number of the arguments in the server group to add
- also.

Changed general option ca_certs to ca-certs.

This involved changes to the class commands and the instance commands.

We did not change: propertylist or namespace.