Remove --sort option from class and instance  commands
We had an optional --sort on a commands such as enumerate references,
associations for classes and instances.

Since there is no natural order(the server has no ordering requirements),
I concluded that there is no reason to make it an option. Removing the
option I set the sort to True so that the returned classes, classnames,
instances, and instancenames are always sorted before being displayed.