Fixes issue #258, #259, table output of instance names
the instance enum with -o and table output_format was just outputting
the repr output of the keybindings.  Replaced this with a table preparer that

1. Includes separating the class from the keybindings and putting into
its own table entry. The table now shows host, namespace, class,
keybindings.

2. Formatting the keybindings based on to_wbemuri() and further, if the
result is greater than the width of the table we want (terminal width)
reformatting the keybindings into multiple rows.

Adds test for the new format.

Fixes issues from the change of --noverify

Fixes a couple of issues in the documentation with the change yesterday
from --noverify to --no-verify