Fixed issues in output format documentation
We had two issues in the output formatting for CIM objects:

1. The repr format was docmented, implemented but somewhere we removed
it from the list of available formats. Added it back.

2. The txt format was incomplete.  Set it up. It essentially just
outputs the __str__ for each object.

3. The xml, repr, txt formats had not been tested.  Added tests for the
class command group and TODO for qualdecl and instance tests.

4. The tree format showed up in the list but we never did implement it
that way. It was implemented as the only output for the class tree
command. Removed it from list of output formats. Removed it from the
list of possible formats since, for now, it only shows up the one place.

Fixed documentation to reflect the above and further explain the cases
where the output format general option is honored and ignored.

We added this explanation to the output_formats section

Fix table in doc. Changed the example with table output for connection
list in the doc to show the current correct format.