Karls Changes to Andy's PR/
This commit extends the change of the tests to use innows to all of the
cli tests instance, server, connection

We did not change the help test in general since that is for the top
level help.  It is easy to change and worth maintaining exact comparison
with that file so we are sure to review all changes.

On the additions I made, for each line involving an option I included
the option definition and at least part of the first line of the text.

NOTE: I make one code change in common because I noticed some confusion
during testing. Removed an outer try block and changed the wording
slightly.

We also made a couple of wording changes to the top line for some of the
help text.

Added new  file in tests that defines common help lines.