Fix issue #275 correct documentation/help for certfile, keyfile
The documentation and help were wrong.  I took what was
defined in pywbem documentation for the certfile and keyfile parameters
and put it into the help for the --certfile and --keyfile and the
general documentation for the same.

Added code to refuse keyfile if no certfile and added test for same.