Changed name subcommand to command
Eliminated all use of subcommand in the help, code, and documentation.

Everything is either a command-group or a command.

Whether we change to object/action or leave it like it is, this makes
the whole thing much simpler.

The only subcommand usage in pywbemcli now is a click functtion
invokde_subcommand that probably started the whole command/subcommand
ruckus.

Changed the name of the doc file pywbemclisubcommands.rst to
pywbemclicommands.rst

Note this involve changes to doc, pywbemcli code, and the tests.