Fixes issue #271 error with class enumerate
Fixes issue # 271.
c
This was an error introduced with the table output of enumerate class
with the -o option.

The error was introduced yesterday when we extended the print_paths as
table function for better table output for instances.  We broke the
class output.

Surpise, there was no test for this so particular set of comnand and
args the tests did not catch it.