Fixes issue #261, inconsistent connection commands
The changes for the connection add, delete,
select, test are:

1. Always use the NAME positional argument to name the target connection
(instead of the --name option) for connection add, select, delete, show.
For add NAME is required. For the others it is optional and we help them
chose with selection list.

2. Change the connection select so that the result is persistent and modify the startup behavior so that if there is no --mock-server, --name, --server general
options, pywbemcli tries to use the default name from the
repository file. This means that there are two pieces of state when
pywbemcli is active (current and default connections)

3. Change the show and list so that we display both connections from the
connection file and the current connection if it is not in the file
(ie. if it was just defined with --name or --mock-server but not savedi.
Any connection that is current but has not been saved has the name
"not-saved" and must have the state current since it will disappear if
something else is selected.

4. Change the mechanism that determines which connection to show so the
options are:
  a. provide NAME argument and it finds that name.
  b. No name. If there is a current connection, it shows that
     connection. otherwise it gives a list of the connections in the file to
     use as a selection
  c. use "?" to specifically show the list of all servers including
     any current that is not saved.

5. Extend add with a --name so that we can use an existing names as
the basis for a new name. If the --name is used, that connection
definition is used as the data for the new connection with the name
argument.

6. For the moment we ignored the test and it still uses the current
connection.

7. Made minimal changes to docs/pywbemclicommands.rst so we will have to
   refactor that when the other PRs catch up.

NOTE: This clarifies two forms of state:

  a. default connection - a connection defined in the file and selected
     to be the default with the connection select command.
  b. The current connection, which is the connection to which the
     connection that is currently open (WBEMConnection established)

We show both of these states in the list ouput for each conneciton in
the list. These may be the same connection or different.

pywbemcli -s http://blah     Only has a current connection