Correct an issue found in invokemethod with array parameters.
1. pywbemcli was ignoring all but the last input  parameter.
2. The output formatting was too simplistic and ignored array
   values

Corrects the issues, extends the testing by extending  all_types mof
to add a method that has IN,OUT parameters for many CIM types
and creates a mock invokemethod for this method that returns all
parameters input.

Adds tests for the different CIM Parameter types.

Adds example of invokemethod to documentation.