Fixes issue #313, modify command descriptions to use sections
Modifies all of the command descriptions to a new level of
sections so that they can be seen in index, toc, and also
references.

Note that I changed text on some of the early ones but for most of them
I simply changed the bullets to sections so we can merge more easily
with the other changes occuring simultaneously.