Consolidated pywbem_os_setup.sh script
Details:

The root cause of having to have a local version of pywbem_os_setup.sh
was that the Makefile did not invoke pywbem_os_setup.sh/bat for
development, reasoning that pywbwemtools itself did not have the development
requirements of pywbem. However, on Python 3, the 'pylint' package (via
'astroid') needs the 'typed-ast' package, which depends on the OS-level
packages libcrypt-devel and python3-devel (at least on CygWin). These
OS-level packages happen to also be used by pywbem for development. For
reasons not investigated, this setup worked on all platforms except CygWin.

The fix is to actually invoke pywbem_os_setup.sh/bat for development.

The installation descriptions in the docs and readme files were updated
accordingly.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>