Fix issues # 337 and # 338 Modify command handling to use general options.
Modified the structure to allow use of the general options within an
interactive session.

1. Removed some of the properties (timestats, use_pull, pull_max_cnt,
log, verbose) from PywbemServer class. These are maintained only in the
context and therefore, when set, they apply to all commands, not any
particular pywbem server. They were always accessed from the context
rather than the server so should never have been in the server.

2. Added property setters to PywbemServer class so can individually
set properties.

3. Modified no_verify option. Modified no-verify option to allow 3 values
 (None, True, False). This allows the code to determine if the
option was defined.

4. Reorganized pywbemcli.py slightly to do function cli input argument
processing on all calls, startup and command calls. This produces
resolve_... versions of each of the input arguments so that the code
can determine if the argument was actually input or the default was set.

5. Changed the name of the no_verify option to verify because the order
of processing in the click option that handles '--a/--no_a' is such that
we would have either had to define it as '--no-a/--a' or invert the
argument in processing. Besides this actually seems more logical instead
of the double negative (i.e. is no-verify True or false). Note that the
alternative would have been to make it a choice argument (choice
("true", "false"). We had to do something so that we could determine if
the argument was input or not and the is-flag does not do that.

6. Remove the block on overwrite of connection definitions with
'connection save'.