Change 'subcmd' to 'command' and command group in tests and code (#365)
* Change subcmd to command in tests and code

We have decided on the names command and command group to define the
components of the pywbemcli cmd line.  However they were all named
subcmds in the tests.

The word subcmd should not occur any more.

This just changes the text and the key names in the test dictionaries.

It does not change code.

Also changed the test names from test_*_subcmd.py to test_*_cmd.py

* Remove change that was not supposed to be in this pr

Removed the edit to fix issue #363.