Fixes issue # 360 server connection vs connection show
This is one proposal to just remove this command since it is a duplicate
of the connection show command.

This is just a proposal at this point.  Some day we will have a command
that wants to show a lot of information about the server but that will
be more than just the connection information.

Suggestions welcome