Fixes #572: Add table output for connection show
This pr adds an optional table output for connection show.

As part of that change it expands the fold_string common
method to allow other optional arguments.  This includes folding
lists of strings,adding options for when fold occurs, and
how the lists are folded.  Adds test for fold_string

Fixes issue in test_pywbemserver with extra test output

Changes some test documentation in test_common.py to correctly
reflect the functions being tested.

Adds this change to changes.rst