Fixes Issue # 354, Command-specific default output format & improve output format validation
1. Set default for output format general option to None.

2. Implement testing of each command for allowed output formats
for that command.

  a. Create a function in common.py (validate_output_format)
whose parameters are  the cmdline input format,  the allowed
format types (CIM, TABLE), and the default output_format
if the cmd line output_format is the default (None).  If the function
is called with a output_format that is not
allowed it generates an exception. Otherwise it returns the
output_format to be used for the command.

  b. Add this function to each command action function that
returns data so that each command is tested against the
cmd line output_format for valid output format type and also
provides its default if the general option is None.

Note that this function call is not added to commands like
instance delete that return nothing.

3. Add tests for invalid format for specific commands

4. Adds a test of the new common function.

4. Modify the documentation to include a table of allowed output
formats for each command.