Fixes issue #349 consistency of prompting (#374)
* Fixes issue #349 use of --interactive option

1. Modified code to use ? as a wild card for  the key component of
wbem uri to kick off the selection process.  This modifies all the
following instance commands: get, delete, modify, references, associatiors,
invokemethod.

2. Modified the instance tests and the doc on the instance get, delete,
associations, and references, etc. to reflect this new way of specifying
the INSTANCENAME.

3. I think that the connection show is correct now in that it uses
the current connection if name not provided and uses ? if the user wants
the prompted list.

4. Modified the documentaton for instance get, delete, modify,
references, associators, invokemethod.  I shortened the test in each
command in favor of pointing them to the general section on the
INSTANCENAME in features.

This turns out to be a  really good solution, much better than the
--interactive option.

NOTE: I fixed one ref definition error in pywbemclicmdline.rst

* Fix Andy's Comments

1. Refactored the description of INSTANCENAME argument to only reference
a general section from pywbemclicommands.rst and create the complete
description of both the complete WBEM_URL and selection into a section
in features with subsections for WBEM URI and for selection.

2. I think I fixed the other issues also