Allow turning off --verbose in interactive mode by changing the long definition name to --verbose/--noverbose and the default for verbose to None on the cli function call.
This PR also adds a display when verbose is set that is turning
output to be worthwhile. It displays the current command and its
parameters when --verbose is set just before the command is called. Note
that this only displays that command, not the command group name

Finally it adds a level counter to a debug display of the tree of
click contexts.  NOTE : This is debug ONLY