Fix issue # 394 - Aborts session in interactive mode
We agreed that we limit the issues that cause abort of the session to
just errors in creating

Removed the aborts on all errors except mock-server MOF or python script
errors in the files themselves. Even nonexistent or misnamed files do
not cause an abort.

Adds tests for both the Abort from interactive and the cmds that do not
abort in interactive mode.