Extend coverage by adding tests for areas that are currently shown to not have any coverage in coverage_report
Adds tests to increase coverage in

1. pywbemcli.py - Adds tests to catch a number of the documented
exceptions in pywbemcli.

2. Add tests for password prompt

3. Added tests for other exceptions I noted were not tested with the
coverage report.

NOTE: All this and I picked up about 1%. Note that in the last few weeks
we went from 200+ tests to over 400 and only picked up about 10%. The
next few percent are going to be difficult since each little piece of
code is generally a whole test

Code changes.

I made two code changes here because they are very small
and because they had a direct impact on the tests I was adding:

1. Found an error that only affects our tests where we want to compile
some mock script during startup.  We were using the cli argument
mock_server internally and should have been using the variable
resolved_mock_server so that the startup compiled script was shown as a
file that gets added to the mock_server. Now these files are compiled
during startup and not included as part of the mock_server.

2. We had left one place in the code where a script compile exception
did not cause an abort (pywbemcli.py, line 414.  I changed that.

3. pywbemcli line 435 testing choices for use_pull.  We did not need to
do this test because the option handler does it for us. That was dead code so
I just removed it.