Fixes issue #255 Reorg doc for multiple tools
Reorganize the documentation to allow for multiple tools.  This pr moves
all of the pywbemcli specific documentation to a new subdirectory and
creates an toctree (index.py) for that documentation so that the pywbemcli
documentation all becomes section 2 of the documentation.

Add the :hidden: attribute to the pywbemcli toctree so that the TOC for
section 2 does not show up in the page.

We were having problems with display of the TOC for chapter 2 in the
sidebar (it showed the other components of the sidebar but not the Table
of Contents when Chapter 2 was selected.  Resolved this by adding a
new Sphinx package (fulltoc).

This modifies the side bar to always show all Chapters and also show the
contents of the selected chapter.

Move mock support chapter to section of pywbemcli command

The mock support generally applies today to the pywbemcli command.

Therefore this chapter should be a section of the pywbemcli command.

If in the future we mock other tools, we can reorganize the mock
documentation.