Fixes issue #443, ordering of lists of commands in help
Changes the ordering of the lists of commands in help from the
alphabetic to the order that the commands in each group are  defined
for command groups except for the top level group.

For the top level group, we leave the ordered listing
of commands in the help indeterminate except to move connecton, help,
and repl to the bottom of the list.