Fixes issue # 443 - Filter of classes
Adds the capability to filter classes returned from a number of
operations by

1. Whether they are an association or now.
2. Whether they are indications or not
3. Whether the class is experimental or not

If the option for a particular filter is not defined, the filtering
for that filter is ignored.

filters:

1. --association - return only association classes
2. --no-association - return only classes that are not associations
3. --indication - Return only classes that are indications
4. --no-indication - Return only classes that are not indications
5. --experimental - return classes with experimental qualifier
6. --no-experimental - return classes that do not have experimental

The commands that do this filtering are:

1. class enumerate
2. class find

TODO:

1. Extend to instance count

Adds a new model to tests for qualifier filter tests

This pr adds the qualifier_filter_model.mof to tests/unit. This is based
on the simple_assoc_model but adds the experimental qualifier and
some experimental classes.  It adds a number of tests of various
combinations of the qualifier filter options.