Hide diagnostic display
Hide a diagnostic display that was being output in --verbose mode but
was really a developer only display.

We did this by adding a test for an undocumented env variable in
combination with the --verbose mode so that even the test for the
env-variable only occurs when in verbose mode.  The alternative was to
remove it and we may still need it for some development inspection.

The code being hidden displays the full click context for each command
and is of no use to anybody except developers the next level of
commands, etc.  In reality it is not of any use in debugging existing
commands once they have been proven to work. But we may need it to sort
out some of the issues in stdin processing and more complex commands in
the future.