pick_from_list bypass user request if only one choice
Modify the pick_from_list() that normally issues a list of choices for the
user to pick one without issuing the pick list when there is only a
single choice and return that choice.  This bypasses the user having
to pick from a choice of one when there is only one choice.

Adds test for this and a change to the documentation to document this
behavior.