KS/#2062a - Add NoCaseList function
Adds a simple NoCaseList function to _common.py along with tests.

This function will be used by the association shrub code but it probably
more general than just that class.

The function allows doing "for a in b" with case insensitive strings and
is used for tests like if role in roles, if clasname in classnames.