Refactor tree display code
Refactor the code in the _displaytree.py module to make it more
comprehensible, faster, and correct an outstanding issue where we
were not always displaying the tree in the same order.

This also adds tests for boundary conditions.

This pr does not have a defined issue since it does not change
external behavior.