Remove use of NocaseDict from class tree command
Remove _display_tree dependency on pywbem NocaseDict

This replaces NocaseDict with a PYPI package pydicti that contains both
a Nocase dictionary and a Ordered Nocase dictionary.

Note that using the ordered dictionary clears another issue in
pywbemcli, #135, Tree output inconsistent ordering in that the
dictionary used throughout this process is not ordered.