Fixes issue #510 - Order of display of commands in help inconsistent
This pr changes the order of executing the __init__ super class call
in the PywbemcliGroup __init__() since click actually hard codes
the {} in in the Group class __init__ and only executes the change
if the python version does not order dictionaries.  Sadly that means
that a number of the slack and click discussions that propose the
original code do not actually work but that is their problem

Added tests for class and instance groups to specifically insure that
the order of the commands displayed in the group help is the same
for different python versions.

Update cmdshelp.rst