Hide the text output from a number of tests.
Since changing the use of tools like simplified_test_function to allow
capturing text output (i.e. using the capsys fixture is signicantly
difficult, we found a simple mechanism, simply mock the click.echo so
that the output is never generated.  We added mock of click.echo
to the mock we already had for click.confirm, click.prompt, etc.

Changed the tests in common that output text so that the test output is
significantly cleaner now.