Fixes ks/#556 brief and full connection list table display
Modify connection list table to have a brief table form
that shows only name, server, and mock_server columns and
an option --full for a display of all of the fields (except
password and ca-certs which were never displayed).

Remove the option that hides empty columns of the table so
that all columns in the defined list show even if there is no
data in the column.

Added tests for the new table format.

Modified the docs/pywbemcli/commands.rst to show this option
and show the tables with and without the option.