Fixes Issue #595 document search in interactive mode
This pr documents an option of the prompt-toolkit to allow incrementally
searching the history file in interactive mode.

The search options is initiated by <CTRL-r> and does an incremental
search based on a string entered after the <CTRL-r> returning the last
command in the command history that contains the the string.  The
string may be modified or <UP_ARROW>, <DOWN-ARROW> may will find
other instances of the string.

It changes the repl help text and the doc files to provide documentation on
this search option and a simple example.

There are no code changes and no tests for the search command.

We added some doc index entries also.