Fixes issue # 594 Simplify output and implement TEXT ooutput group
Implements a new output format group (TEXT) and modifies two commands to
use this group (server interop, and server brand). This removes the
table option for these commands.

It also implements the text output format as an optional format for
namespaces so that the namespace list appears as a comma-separated
line of text.  In this case the default is table and the optional
choice is text.

Modified the definition fo possible output formats to use a
OrderedDict as the basis for defining the list of formats for each
group and the corresponding default keywords.