Issue #504, Changes to pywbemcli required for pywbem 1.0.0 compatibility.
A number of changes were required to make pywbemcli compatible with both
pywbem pre version 1.0 and the incompatible differences in pywbem 1.0.0.

Generally these are issues in the tests and not in pywbemcli itself.

Changed the tests so that we create variables PYWBEM_0 and PYWBEM_1 in
a common file and use these variables to modify tests.

Modified a number of tests to account for the fact that the returned
url for many tests now includes the port in addition to the host.

Modified the invokemethod test code to create the user-defined test
provider required by pywbem version 1
Since the pywbem 1 invokemethod server responder is completely different
that the pywbem 0 version, two different InvokeMethod test driver python
files are for the invoke method tests are defined.  This applies to both the
simple_mock_invokemethod.py code and the all_types_method_mock.py
code

1. Changed name of original file to append _V0 to the name

2 Created corresponding _V1.py as the user-defined provider for
InvokeMethod tests.

3. Added if statement to each test file that tests invokemethod to use the
correct file.

4. Created new test for testing url invalid port for pywbem 1.