Fixes issue #606 - Deal with server that has no class operations
1. Adds extra logic for a server that returns NOT_SUPPORTED for an
EnumerateClasses operation on the default namespace.

2. Also adds support for testing for existence of  the pull operations
as an option (--test-pull) to the `connection test` command.

Tests added for the pull operations option.  Note that this is only for
the good response since pywbemcli has no easy way to turn off the
pull.  Marked that with TODO: Future.

Changes text of documentation connection test command