Fixes #606: Fix connection test cmd (#615)
* Fixes issue #606, connection test too limited.

Expand the connection test command so that it can test a
server where the class requests are not supported by executing
an EnumerateInstances request against a common class.

Add tests for the pull operations as an option (--test-pull) of the
connection test command.

* Fixes issue #606 - Deal with server that has no class operations

1. Adds extra logic for a server that returns NOT_SUPPORTED for an
EnumerateClasses operation on the default namespace.

2. Also adds support for testing for existence of  the pull operations
as an option (--test-pull) to the `connection test` command.

Tests added for the pull operations option.  Note that this is only for
the good response since pywbemcli has no easy way to turn off the
pull.  Marked that with TODO: Future.

Changes text of documentation connection test command