Fixes issue #637-MOF compiler-display-compile-errors
Modify code to display compile errors both for pywbem 1 and previous
versions of pywbem.

For pywbem 1 it displays the exceptions.  For earlier versions, it
notes that the error preceeds the exception text.

Note that for pywbem 1 the compile error text and exception message
are routed to stderr whereas before the compile error message was
routed to stdout but the exception message to stderr.

NOTE: We now catch all Error exceptions to keep a common exception
name for all versions.

Modified second use of pywbem version to use full version parse.

Added requirement for packaging module because added parse

Fixed issue where the minimum_constraints for packaging was 19.2 and
pywbem, etc. are using 17.0 so I went with 17.0

Found minor pywbem issue where OpenExecQuery tests request parameters before
testing valid namespace.  This is not a real issue except that one of
our pywbemcli tests fails the bad namespace test.  Filed issue #2313 on
pywbem and disabled test with comment about issue.