Fixes issue # 663, connection file load failure
Fixes issues where connection file load from yaml not type validated.
Also extends exception messages to clarify problem in each case.

The original problem that caused this was a side issue, somewhere
invalid data had gotten into a pywbemcli_connection_definitions.yaml
file (ca_certs had a list type). The old pywbem did not catch this
but the spec was always string and not list of strings.

However, it pointed out that we were not doing enough testing in
the Pywbem_Server class constructor and methods.