Fixes issue #612 - Create new groupi named  "profile"
This PR  creates a new pywbemcli command group "profile" and moves the
profile oriented commands (profiles to list, and centralinsts) to that
group.  This PR is simply to move the existing commands definitions for
the server profiles and server centralinsts commands from the server
group to the profile group.

It does not modify the function of the commands except it adds to output
table the options selected.

Adds new test file for the profile group and entry in changes.rst using
the tests from test_server_cmd.py

Adds comment to changes.rst incompatibilities because the cmd line
interface changed to remove old commands.

NOTE: There is question in the corresponding issue about changing
from an option to argument for the option name.  We did not do that and
I suggest we try this before deciding.