Fixes issue #657 - Losing output instances in shrub when host name used
Not all the instances were produced on output when the host name was
included in the instancename.

In general, the host name is of no value so we remove it at the
beginning of the method.  The call must be for the current host since
pywbemcli cannot dynamically connect to any other host.

Took the opportunity to clean up some other code.

1. Refactored the AssociationShrub by

   a. Moving one group of code to a new method

   b. Replacing the local sort_instances and sore_instancenames
      with the general sort_cimobjects

2. Add documentation to AssociationShrub methods.

3. Replaced the context parameter in AssociationShrub with conn since
   only the conn is used from the context.

4. Fixed minor documentation errors in common.