fix issue #678-add-class-filters
Adds the --deprecated filter to the same commands as the other filters
and extends the filter search to that the qualifiers are inspected
for all class elements defined in the scope for the defined filter.
Thus, an experimental qualifier on a property, method, or parameter
will now be equal to the same qualifier at the class level.

Modifies the filter options definition in each command to group the
qualifier filters so that they can be applied to each command as
a single line and the grouping is defined in _common_options.py

NOTE: The logic for multiple filters is effectively anding of
the conditions so that --association and --indication is for both
the association and indication filter and --association --no-indication
finds classes that have the association qualifier but not the
indication qualifier.

The code was refactored to make extension to other filters easier in
the future.

Added more options to the test mof qualifierfilters.mof and extended
the tests.