Fix issue #680 - tests lose default connection file.
This fixes an issue where the default connection file was not being
restored in all cases when tests were run. It also cleans up the
nameing of the paths to be saved and restored to be more readable.

Note that this is a test change only.  It does not affect our CI tests
but could affect a user who had a data in the default connections file
and ran the pywbemtools test suite.