Fixes issue #694-pull-max-cnt
This change moves pull-max-cnt into the pywbem_server so it becomes
a part of the persistent attributes of a server rather than just
a transient parameter.

This modifies the class PywbemServer to add pull_max_cnt and also
modifies the server setup.  Note that if the user has not
specified a specific pull-max-cnt, the connection reports show this
emtpy since we use the default value and do not want to put that
into the persistent store.

It also corrects a couple of issues that for some reason hit us this
time (table layout in the qualifier enumerate) and a report output
in test_instances_cmd.py

Adds item to changes.rst