Fix test issues with use of pywbem 1.0.0. This includes:
This pr is based on PR #711, test method invocations.

NOTE: It was tested against pywbem code with the OpenQuery changes

1. Getting NocaseDict  User warning message in some tests that causes tests to fail. Fixed
by changing to use innows instead of lines as test so that the
warning messages are ignored.

2. Fix issues where for some reason the CIMError message has
changed from NOT_FOUND to INVALID_CLASS. This is one single test
"instance enumerate" in test_instances_cmd.py.  No idea why that changed.

3. Qualifier tests changed number of qualifiers in mock_simple_model.mof
Adds new static qualifier to the mock file

4. Modified the invoke_method test method provider. There is an issue with
returning the params as a NocaseDict. I modified the
code to return a list of CIMParameter items. Now passes tests.

5. Added tests for --deprecation-warnings general options.